refactor: break down the code to separate components
